backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-generativeai==0.8.4
pydantic==2.10.3
python-multipart==0.0.19
python-dotenv==1.0.1
